backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
prometheus-client==0.21.0
aiohttp==3.10.5
SQLAlchemy==2.0.36
alembic==1.13.3
